Fix crash when WangColor probability is int

+++ lib/tile_set/wang_color.dart
@@ -10,7 +10,7 @@ class WangColor {
   WangColor.fromJson(Map<String, dynamic> json) {
     color = json['color'];
     name = json['name'];
-    probability = json['probability'];
+    probability = json['probability'].toDouble();
     tile = json['tile'];
     if (json['properties'] != null) {
       properties = <Property>[];
